requirements.txt
openai
python-dotenv
faiss-cpu
tqdm
gitpython
watchdog